Add complete scripts to install Kubernetes v1.5.3

+++ tools/configure-node.sh
@@ -4,17 +4,20 @@ set -eu
 export LC_ALL=C
 
 ROOT=$(dirname "${BASH_SOURCE}")
-source ${ROOT}/env.sh
-source ${ROOT}/util.sh
+source ${ROOT}/default-env.sh
+source ${ROOT}/utils.sh
 
 TYPE=$1
 TARGET_IP=$2
-TARGET=${USERNAME}@${TARGET_IP}
+TARGET=${NODE_USERNAME}@${TARGET_IP}
 
 kube-ssh "${TARGET}" "sudo rm -rf ${KUBE_TEMP}"
 kube-ssh "${TARGET}" "mkdir -p ${KUBE_TEMP}"
 
-kube-scp ${TARGET} "${ROOT}/env.sh" "${KUBE_TEMP}"
+kube-scp ${TARGET} "${ROOT}/default-env.sh" "${KUBE_TEMP}"
+if [ -f "${ROOT}/env.sh" ]; then
+  kube-scp ${TARGET} "${ROOT}/env.sh" "${KUBE_TEMP}"
+fi
 kube-scp ${TARGET} "${ROOT}/common/*" "${KUBE_TEMP}"
 kube-scp ${TARGET} "${ROOT}/${TYPE}/*" "${KUBE_TEMP}"
